+++ bukkit/src/main/java/net/momirealms/craftengine/bukkit/entity/data/MobData.java
@@ -1,7 +1,6 @@
-package net.momirealms.craftengine.bukkit.entity;
+package net.momirealms.craftengine.bukkit.entity.data;
 
 public class MobData<T> extends LivingEntityData<T> {
-
     public static final MobData<Byte> MobFlags = new MobData<>(15, EntityDataValue.Serializers$BYTE, (byte) 0);
 
     public MobData(int id, Object serializer, T defaultValue) {
