+++ gradle.properties
@@ -31,7 +31,7 @@ cloud_minecraft_extras_version=2.0.0-beta.10
 boosted_yaml_version=1.3.7
 bstats_version=3.0.2
 caffeine_version=3.1.8
-rtag_version=1.5.9
+rtag_version=1.5.10
 placeholder_api_version=2.11.6
 vault_version=1.7
 guava_version=33.3.1-jre
@@ -49,7 +49,7 @@ mojang_brigadier_version=1.0.18
 byte_buddy_version=1.15.11
 snake_yaml_version=2.3
 anti_grief_version=0.13
-nms_helper_version=0.19
+nms_helper_version=0.21
 # Ignite Dependencies
 mixinextras_version=0.4.1
 mixin_version=0.15.2+mixin.0.8.7
